jobappman/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-rest-auth==2.2.4
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2022.1
sqlparse==0.4.2
